I renamed some of the fields to be formatted/non formatted in db. + More
I need to figure out how to search by a range of shifts. I'd like to make it a generic function that could be applied to any data type in the future. Such as the Between field I added to the filter object. But I need to achieve something like SELECT * FROM shifts WHERE start BETWEEN {then} AND {now} OR end BETWEEN {then} AND {now}; because if  I only filtered by one of those fields i might not get a complete list of shifts.