Added handling for date_from and date_to
The way this works is you would add the tag "range" to a field in Golang (the field in PostgreSQL must be convertable to a date). And when the url params are present it will add a greater than and/or a less than to the query for the fields. But both the starting and ending field do need to be defined in a struct.