Removed all timezones. See notes.
Unless I'm working exclusively with UTC and then converting it back to the right time on the client side (Which is also weird because the start and end times of a week are midnight in utc, but are 7pm in CDT, so I'd have to make a note to offset the start and end times in utc a certain ammount).

All in all, this is a test; and while timezones are relevant, it seems better to me to just store them as they are for now and actually make the software work and then come back to it later if i have time.